+++ src/functions/generator.ts
@@ -1,4 +1,7 @@
-import { getTag } from '../utils/to-string.js'
+import {
+  hasGeneratorFunctionBrand,
+  hasGeneratorShape,
+} from '../utils/brands.js'
 
 /**
  * Check if value is a generator function
@@ -10,8 +13,10 @@ import { getTag } from '../utils/to-string.js'
  * isGeneratorFunction(function*() {})  // true
  * isGeneratorFunction(function() {})   // false
  */
-export function isGeneratorFunction(value: unknown): value is GeneratorFunction {
-  return getTag(value) === 'GeneratorFunction'
+export function isGeneratorFunction(
+  value: unknown
+): value is (...args: never[]) => Generator<unknown, unknown, unknown> {
+  return hasGeneratorFunctionBrand(value)
 }
 
 /**
@@ -24,6 +29,8 @@ export function isGeneratorFunction(value: unknown): value is GeneratorFunction 
  * isGenerator((function*() {})())      // true (generator instance)
  * isGenerator(function*() {})          // false (generator function)
  */
-export function isGenerator(value: unknown): value is Generator {
-  return getTag(value) === 'Generator'
+export function isGenerator(
+  value: unknown
+): value is Generator<unknown, unknown, unknown> {
+  return hasGeneratorShape(value)
 }