+++ src/objects/array.ts
@@ -9,6 +9,10 @@
  * isArray([1, 2, 3])        // true
  * isArray('not array')      // false
  */
-export function isArray<T = unknown>(value: unknown): value is T[] {
-  return Array.isArray(value)
+export function isArray(value: unknown): value is unknown[] {
+  try {
+    return Array.isArray(value)
+  } catch {
+    return false
+  }
 }